More Packages and Home with navigation
Beside the navigation packages and eslint, it was implanted a molecular Card and a HomeList organism to reuse on the home page and some others. A styles folder was created, where colors and fonts were added, and also metrics with logic to turn pages responsive for different screen size devices. Also,